requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn==1.0.2
statsmodels==0.13.5
tensorflow-cpu==2.11.0
tensorflow-probability==0.19.0